+++ NEWS.md
@@ -1,9 +1,9 @@
 # proporz (development version)
 
 * add new feature: winner take one method for `biproporz()` and `pukelsheim` (#10)
+* allow non-integer vote counts
 * round divisors to as few digits as possible
-* added more test for real life data
-* bugfixes and updated documentation
+* fixed bugs and added more real data tests
 
 # proporz 1.4.0